refactor(capsule): address PR review — simplify merge, extract tests, remove double KV read
- Replace imperative final_prompt construction with declarative Vec::join
- Move 36 tests from lib.rs to separate tests.rs module (lib.rs: 1060→530 LOC)
- Inline provider derivation in handle_identity_response, remove active_provider_id()
  to eliminate redundant active_llm_topic() call
